+++ FXLabel/FXLabel.m
@@ -1,7 +1,7 @@
 //
 //  FXLabel.m
 //
-//  Version 1.2
+//  Version 1.3
 //
 //  Created by Nick Lockwood on 20/08/2011.
 //  Copyright 2011 Charcoal Design. All rights reserved.
@@ -31,22 +31,38 @@
 #import "FXLabel.h"
 
 
+@interface FXLabel ()
+
+@property (nonatomic, assign) NSUInteger minSamples;
+@property (nonatomic, assign) NSUInteger maxSamples;
+
+@end
+
+
 @implementation FXLabel
 
 @synthesize shadowBlur;
 @synthesize innerShadowOffset;
 @synthesize innerShadowColor;
-@synthesize gradientStartColor;
-@synthesize gradientEndColor;
+@synthesize gradientColors;
 @synthesize gradientStartPoint;
 @synthesize gradientEndPoint;
-@synthesize oversample;
-
+@synthesize oversampling;
+@synthesize minSamples;
+@synthesize maxSamples;
+@synthesize textInsets;
 
 - (void)setDefaults
 {
     gradientStartPoint = CGPointMake(0.5f, 0.0f);
     gradientEndPoint = CGPointMake(0.5f, 0.75f);
+    minSamples = maxSamples = 1;
+    if ([UIScreen instancesRespondToSelector:@selector(scale)])
+    {
+        minSamples = [UIScreen mainScreen].scale;
+        maxSamples = 32;
+    }
+    oversampling = minSamples;
 }
 
 - (id)initWithFrame:(CGRect)frame
@@ -68,15 +84,6 @@
     return self;
 }
 
-- (void)setShadowBlur:(CGFloat)blur
-{
-    if (shadowBlur != blur)
-    {
-        shadowBlur = blur;
-        [self setNeedsDisplay];
-    }
-}
-
 - (void)setInnerShadowOffset:(CGSize)offset
 {
     if (!CGSizeEqualToSize(innerShadowOffset, offset))
@@ -96,31 +103,81 @@
     }
 }
 
+- (UIColor *)gradientStartColor
+{
+    return [gradientColors count]? [gradientColors objectAtIndex:0]: nil;
+}
+
 - (void)setGradientStartColor:(UIColor *)color
 {
-    if (gradientStartColor != color)
+    if (color == nil)
     {
-        [gradientStartColor release];
-        gradientStartColor = [color retain];
-        [self setNeedsDisplay];
+        self.gradientColors = nil;
     }
+    else if ([gradientColors count] < 2)
+    {
+        self.gradientColors = [NSArray arrayWithObjects:color, color, nil];
+    }
+    else if ([gradientColors objectAtIndex:0] != color)
+    {
+        [gradientColors autorelease];
+        NSMutableArray *colors = [gradientColors mutableCopy];
+        [colors replaceObjectAtIndex:0 withObject:color];
+        gradientColors = [colors copy];
+        [colors release];
+    }
+}
+
+- (UIColor *)gradientEndColor
+{
+    return [gradientColors lastObject];
 }
 
 - (void)setGradientEndColor:(UIColor *)color
 {
-    if (gradientEndColor != color)
+    if (color == nil)
     {
-        [gradientEndColor release];
-        gradientEndColor = [color retain];
+        self.gradientColors = nil;
+    }
+    else if ([gradientColors count] < 2)
+    {
+        self.gradientColors = [NSArray arrayWithObjects:color, color, nil];
+    }
+    else if ([gradientColors lastObject] != color)
+    {
+        [gradientColors autorelease];
+        NSMutableArray *colors = [gradientColors mutableCopy];
+        [colors replaceObjectAtIndex:[colors count] - 1 withObject:color];
+        gradientColors = [colors copy];
+        [colors release];
+    }
+}
+
+- (void)setGradientColors:(NSArray *)colors
+{
+    if (gradientColors != colors)
+    {
+        [gradientColors release];
+        gradientColors = [colors copy];
         [self setNeedsDisplay];
     }
 }
 
-- (void)setOversample:(BOOL)over
+- (void)setOversampling:(NSUInteger)samples
 {
-    if (oversample != over)
+    samples = MIN(maxSamples, MAX(minSamples, samples));
+    if (oversampling != samples)
     {
-		oversample = over && [[[UIDevice currentDevice] systemVersion] floatValue] >= 4.0f;
+		oversampling = samples;
+        [self setNeedsDisplay];
+    }
+}
+
+- (void)setTextInsets:(UIEdgeInsets)insets
+{
+    if (!UIEdgeInsetsEqualToEdgeInsets(textInsets, insets))
+    {
+        textInsets = insets;
         [self setNeedsDisplay];
     }
 }
@@ -161,13 +218,18 @@
 
 - (CGColorRef)color:(CGColorRef)a blendedWithColor:(CGColorRef)b
 {
-    CGFloat aRGBA[4];
-    [self getComponents:aRGBA forColor:a];
-    if (aRGBA[3] == 1.0f)
+    CGFloat alpha = CGColorGetAlpha(a);
+    if (alpha == 1.0f)
     {
-        return [UIColor colorWithRed:aRGBA[0] green:aRGBA[1] blue:aRGBA[2] alpha:aRGBA[3]].CGColor;
+        return a;
+    }
+    else if (alpha == 0.0f)
+    {
+        return b;
     }
     
+    CGFloat aRGBA[4];
+    [self getComponents:aRGBA forColor:a];
     CGFloat bRGBA[4];
     [self getComponents:bRGBA forColor:b];
     CGFloat source = aRGBA[3];
@@ -181,16 +243,23 @@
 - (void)drawRect:(CGRect)rect
 {
     //get drawing context
-	if (oversample)
+	if (oversampling > minSamples)
     {
-        UIGraphicsBeginImageContextWithOptions(rect.size, NO, [[UIScreen mainScreen] scale] * 2.0f);
+        UIGraphicsBeginImageContextWithOptions(rect.size, NO, oversampling);
     }
     CGContextRef context = UIGraphicsGetCurrentContext();
+    
+    //apply insets
+    rect = self.bounds;
+    rect.origin.x += textInsets.left;
+    rect.origin.y += textInsets.top;
+    rect.size.width -= (textInsets.left + textInsets.right);
+    rect.size.height -= (textInsets.top + textInsets.bottom);
     
     //get label size
     CGRect textRect = rect;
     CGFloat fontSize = self.font.pointSize;
-    if (self.adjustsFontSizeToFitWidth)
+    if (self.adjustsFontSizeToFitWidth && self.numberOfLines == 1)
     {
         textRect.size = [self.text sizeWithFont:self.font
                                     minFontSize:self.minimumFontSize
@@ -201,28 +270,34 @@
     else
     {
         textRect.size = [self.text sizeWithFont:self.font
-                                       forWidth:rect.size.width
+                              constrainedToSize:rect.size
                                   lineBreakMode:self.lineBreakMode];
     }
     
     //set font
     UIFont *font = [self.font fontWithSize:fontSize];
     
+    //set color
+    UIColor *highlightedColor = self.highlightedTextColor ?: self.textColor;
+    UIColor *textColor = self.highlighted? highlightedColor: self.textColor;
+    textColor = textColor ?: [UIColor clearColor];
+    
     //set position
     switch (self.textAlignment)
     {
         case UITextAlignmentCenter:
         {
-            textRect.origin.x = (rect.size.width - textRect.size.width) / 2.0f;
+            textRect.origin.x = rect.origin.x + (rect.size.width - textRect.size.width) / 2.0f;
             break;
         }
         case UITextAlignmentRight:
         {
-            textRect.origin.x = rect.size.width - textRect.size.width;
+            textRect.origin.x = textRect.origin.x + rect.size.width - textRect.size.width;
             break;
         }
         default:
         {
+            textRect.origin.x = rect.origin.x;
             break;
         }
     }
@@ -232,19 +307,19 @@
         case UIViewContentModeTopLeft:
         case UIViewContentModeTopRight:
         {
-            textRect.origin.y = 0.0f;
+            textRect.origin.y = rect.origin.y;
             break;
         }
         case UIViewContentModeBottom:
         case UIViewContentModeBottomLeft:
         case UIViewContentModeBottomRight:
         {
-            textRect.origin.y = rect.size.height - textRect.size.height;
+            textRect.origin.y = rect.origin.y + rect.size.height - textRect.size.height;
             break;
         }
         default:
         {
-            textRect.origin.y = (rect.size.height - textRect.size.height)/2.0f;
+            textRect.origin.y = rect.origin.y + (rect.size.height - textRect.size.height)/2.0f;
             break;
         }
     }
@@ -257,7 +332,7 @@
     ![self.innerShadowColor isEqual:[UIColor clearColor]] &&
     !CGSizeEqualToSize(innerShadowOffset, CGSizeZero);
     
-    BOOL hasGradient = gradientStartColor && gradientEndColor;
+    BOOL hasGradient = [gradientColors count] > 1;
     
     BOOL needsMask = hasInnerShadow || hasGradient;
     
@@ -280,16 +355,16 @@
     {
         //set up shadow
         CGContextSaveGState(context);
-        CGFloat textAlpha = CGColorGetAlpha(self.textColor.CGColor);
+        CGFloat textAlpha = CGColorGetAlpha(textColor.CGColor);
         CGContextSetShadowWithColor(context, self.shadowOffset, shadowBlur, self.shadowColor.CGColor);
-        [needsMask? [self.shadowColor colorWithAlphaComponent:textAlpha]: self.textColor setFill];
+        [needsMask? [self.shadowColor colorWithAlphaComponent:textAlpha]: textColor setFill];
         [self.text drawInRect:textRect withFont:font lineBreakMode:self.lineBreakMode alignment:self.textAlignment];
         CGContextRestoreGState(context);
     }
     else if (!needsMask)
     {
         //just draw the text
-        [self.textColor setFill];
+        [textColor setFill];
         [self.text drawInRect:textRect withFont:font lineBreakMode:self.lineBreakMode alignment:self.textAlignment];
     }
     
@@ -314,15 +389,18 @@
         
         if (hasGradient)
         {
-            //pre-blend
-            CGColorRef startColor = [self color:gradientStartColor.CGColor blendedWithColor:self.textColor.CGColor];
-            CGColorRef endColor = [self color:gradientEndColor.CGColor blendedWithColor:self.textColor.CGColor];
-            
+            //create array of pre-blended CGColors
+            NSMutableArray *colors = [NSMutableArray arrayWithCapacity:[gradientColors count]];
+            for (UIColor *color in gradientColors)
+            {
+                CGColorRef colorRef = [self color:color.CGColor blendedWithColor:textColor.CGColor];
+                [colors addObject:(id)colorRef];
+            }
+    
             //draw gradient
             CGContextScaleCTM(context, 1.0, -1.0);
             CGContextTranslateCTM(context, 0, -rect.size.height);
-            CFArrayRef colors = (CFArrayRef)[NSArray arrayWithObjects:(id)startColor, (id)endColor, nil];
-            CGGradientRef gradient = CGGradientCreateWithColors(NULL, colors, NULL);
+            CGGradientRef gradient = CGGradientCreateWithColors(NULL, (CFArrayRef)colors, NULL);
             CGPoint startPoint = CGPointMake(textRect.origin.x + gradientStartPoint.x * textRect.size.width,
                                              textRect.origin.y + gradientStartPoint.y * textRect.size.height);
             CGPoint endPoint = CGPointMake(textRect.origin.x + gradientEndPoint.x * textRect.size.width,
@@ -334,7 +412,7 @@
         else
         {
             //fill text
-            [self.textColor setFill];
+            [textColor setFill];
             CGContextFillRect(context, textRect);
         }
         
@@ -343,7 +421,7 @@
         CGImageRelease(alphaMask);
     }
     
-    if (oversample)
+    if (oversampling)
     {
         UIImage *image = UIGraphicsGetImageFromCurrentImageContext();
         UIGraphicsEndImageContext();
@@ -354,8 +432,7 @@
 - (void)dealloc
 {
     [innerShadowColor release];
-    [gradientStartColor release];
-    [gradientEndColor release];
+    [gradientColors release];
     [super dealloc];
 }