+++ Tests/Highlight Color/FXLabelTest/AppDelegate.h
@@ -1,0 +1,19 @@
+//
+//  AppDelegate.h
+//  FXLabelTest
+//
+//  Created by Nick Lockwood on 02/12/2011.
+//  Copyright (c) 2011 Charcoal Design. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@class ViewController;
+
+@interface AppDelegate : UIResponder <UIApplicationDelegate>
+
+@property (strong, nonatomic) UIWindow *window;
+
+@property (strong, nonatomic) ViewController *viewController;
+
+@end