+++ Tests/Highlight Color/FXLabelTest/main.m
@@ -1,0 +1,18 @@
+//
+//  main.m
+//  FXLabelTest
+//
+//  Created by Nick Lockwood on 02/12/2011.
+//  Copyright (c) 2011 Charcoal Design. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+#import "AppDelegate.h"
+
+int main(int argc, char *argv[])
+{
+    @autoreleasepool {
+        return UIApplicationMain(argc, argv, nil, NSStringFromClass([AppDelegate class]));
+    }
+}